requirements.txt
django
django-ninja
psycopg
psycopg2
psycopg2-binary
Pillow
pydantic[email]
celery